decouple fetching from scraping
Introduce the concept of manifests, which are surface-level snapshots
which do not involve any other extra information. To obtain such
information, it must be "gleaned" separately.

In the old design, the construction of a build struct forced at least
one fetch, as we needed to extract the build number and hash from a
bundle. The manifest doesn't require this, permitting construction
before fetching.

Gleaning turns the manifest into a build, which requires a fetch in
order to fetch relevant information. Now though, fetching logic has been
moved out of the processing functions and the assets' contents live
longer, elsewhere, meaning that we get to avoid duplicate fetches and
fetch everything exactly once.

The `Wrecker` struct is used to orchestrate fetching, store the content
of assets, and handle gleaning.