Dump modules' function code
Dump each module's JS when dumping modules. This will serve as the basis
for module diffing in the future. Excitingly, we are able to do this
without copying.

Additionally rename the WebpackChunks dump item to WebpackModules,
because we're only dumping a single set of modules from the main
entrypoint chunk--not all chunks, as one would likely expect from a dump
item named WebpackChunks.