Abstract asset tracking and fetching
Instead of passing an "asset content map" struct around, pass a struct
which hides the details of both finding a certain type of root script
and fetching the content of assets. Offloading these particular concerns
is important as querying assets is central to the process of scraping.

Now, any code that wants to access the entrypoint chunk JS can do so
without caring how to find the script, how to fetch it, or if we've
already fetched it. This interface additionally ensures that we only
ever fetch an asset exactly once.