Unbox dump errors and harmonize nomenclature
dump::DumpError was renamed to dump::DumpWriteError to make it clear
that that error type is used when calling DumpResult::write (renamed
from DumpResult::dump_to). The act of writing a dump to disk should now
be referred to as "writing" instead of "dumping", because "dumping a
dump" to disk is confusing. Now, "dump" solely means extracting
information from an artifact. Meanwhile, actually writing the dump
results to disk is called "writing".

Furthermore, create a proper type for errors that can occur while
dumping and name it dump::DumpError. Now we don't have to type Box<dyn
Error + Send + Sync> everywhere.