store asset content as bytes instead of strings
Not every asset on Discord's CDN is UTF-8 (see: GIFs, PNGs, WEBPs, and
more). We might want to scrape and diff these in the future, so make the
code store asset content as a Vec<u8> instead of a String. This requires
a bit more error plumbing to handle theoretical decode failures, but
it's not too bad.