return u8 slices from asset methods again
We can move the internal logic of the raw_content method into a
freestanding function, which means that we don't have to perform an
exclusive borrow on self for the entire method inside of
preprocessed_content. Now we can avoid unnecessary cloning and return
byte slices again. Guided by assistance from @pluiedev, thanks a ton! :)