make use of the X-Build-ID header
Discord sends the build hash as the X-Build-ID header. We can use this
to detect stale builds with less work and requests made if we store it
in the database, too.

Because we still need the build number, define a new table dedicated to
builds. Builds detected on branches will reference this new table
through a foreign key.

Closes #1.