use Intl.DateTimeFormat
I've verified that the dates being received from the database has the
correct timezone, so we're able to use native web platform date
formatting so that it displays correctly for the user's locale. This
will introduce hydration inconsistencies, but this is seemingly
impossible to workaround (unless we store the user's timezone, but I'd
rather not do that).