separate asset ownership and caching types
Rename Assets to AssetCache, and discharge its responsibility of owning
the actual asset values themselves. This is delegated to a (new) Assets
struct instead, which provides helpers for filtering assets by type (a
common need in the codebase).

The inner Vec is exposed because you often need access to it, and it
doesn't seem like we'd benefit much from hiding the inner interface.

The former coupling didn't make much sense, and made the code somewhat
harder to read as one type was responsible for two seemingly separate
things.