use extension trait instead of newtype for assets
Newtypes in Rust are kinda painful, and we don't really need one here to
encapsulate the idea of "a collection of assets". Since we want to be
able to implement some helper methods that are used quite frequently on
them, though, we can use an extension trait.

This makes me wonder if I should store each asset in a separate
collection keyed by asset type, but then we lose the relative ordering,
which is something that I might care about.