avoid doing cataloging a build more than once
Check if a build is already present in the `detected_builds` table
before doing the work to catalog the total assets of a build. Of course,
this creates an invariant that a build being present in that table means
that we (to the best of our knowledge) already scraped all of a build's
assets.