web: add base build detail view
Navigate to /build/<build_number> or /build/<build_id> to view details
about a build. Some interesting information we show at the moment is
every time this build was detected on a channel (as well as the build it
replaced when it was detected), the build's hash, and its assets.